+++ app/controllers/artist_links_controller.rb
@@ -1,0 +1,2 @@
+class ArtistLinksController < ApplicationController
+end