+++ app/helpers/artist_links_helper.rb
@@ -1,0 +1,2 @@
+module ArtistLinksHelper
+end